Add Hugo
Hugo gives me a framework for a simple, fast and sexy static website
with a slick theme. Right now that's all I want.

I also considered Jekyll and that has the advantage of being ruby, but
hey I'm a software engineer. Right now that decision really was made on
the fact that Hugo has a slicker name and website, with a theme that I
think looks really nice.