Add work and personal social links
Updates social links from template dead links to useful links for people
wanting to connect with me and get to know me a bit.

I considered using the available third column, but there's not really
anything useful to go there. I considered adding email an email but
against putting it in a closing invitation in the main body instead.

I deliberated on the Personal links, whether to have twitter or the
Bible. In the end I decided Bible because I'm more active on there and
it reflects an important and personal part of me, that I like making a
small throw to.