Swap in Twitter for Bible
After thinking about it my online profile is all about new people at the
moment and I'm living a life where like cologne my power in the spirit
is discovered not announced.

I considered leaving the bible out there in the opening for a while but
twitter is something I'm interested in and would appreciate getting a
message from someone through.