Promote Twitter over Facebook
Look I don't mind if people follow me but I'm probably not going to be
adding the people who send me friend requests directly from this page.
My Facebook is a tight list and that's by no mistake.

I considered not having Facebook on this page at all and having the
bible. But the reasons I swapped out the Bible for Twitter still stand,
and I don't really mind people having a little look at Facebook even
sending me a message that I get in messenger from non friends.

The other consideration in this is I shipped and made the changes in the
one commit. This seemed to work in this context for something small. I
don't have fixed thoughts on this as long as it stays thoughtful and
well explained.